+++ packages/jest-validate/src/__tests__/validate.test.js
@@ -266,3 +266,23 @@ test('Repeated types within multiple valid examples are coalesced in error repor
     ),
   ).toThrowErrorMatchingSnapshot();
 });
+
+test('Comments in config JSON using "//" key are not warned', () => {
+  jest.spyOn(console, 'warn').mockImplementation(() => {});
+  const config = {'//': 'a comment'};
+
+  validate(config, {
+    exampleConfig: validConfig,
+  });
+  expect(console.warn).not.toBeCalled();
+
+  console.warn.mockReset();
+
+  validate(config, {
+    exampleConfig: validConfig,
+    recursiveBlacklist: [('myCustomKey': "don't validate this")],
+  });
+  expect(console.warn).not.toBeCalled();
+
+  console.warn.mockRestore();
+});